Create MapControls.js
